GLADDER-10(b): address >=4GB ISOs without growing the ATAPI result struct
media_size (device struct) widens to uint64 so a >=4GB ISO's size is carried
without truncation. The read POSITION, however, must not become a 64-bit byte
offset in hype_atapi_result_t: that struct is a stack local in the AHCI glue and
is size-sensitive -- growing it (either media_offset->uint64 here, or an earlier
MAX_SYNTH_RESPONSE 36->64) reproducibly wedges the inner OVMF's CD enumeration at
INQUIRY on both Fedora and Alpine (root cause unresolved; tracked separately).

So the result reports a 32-bit START SECTOR (media_lba, good to 8TB) instead of a
byte offset, and the glue scales it to a 64-bit byte offset (media_lba * 2048)
at the point of use -- keeping the struct byte-identical to the proven layout.
media_length stays a byte count (one CDB never transfers >4GB). Unit tests cover
the LBA reporting and the >4GB offset arithmetic; Alpine still boots to GRUB.

+++ boot/main.c
@@ -1694,7 +1694,7 @@ static void run_iso_2_test(const hype_vmm_ops_t *ops, hype_vmm_kind_t kind) {
     hype_guest_ram_zero(g_iso_2_guest_code, sizeof(g_iso_2_guest_code));
     hype_guest_ram_zero(g_iso_2_guest_stack, sizeof(g_iso_2_guest_stack));
 
-    hype_atapi_reset(&g_iso_2_atapi, (uint8_t *)(uintptr_t)g_iso_host_phys, (uint32_t)g_iso_size);
+    hype_atapi_reset(&g_iso_2_atapi, (uint8_t *)(uintptr_t)g_iso_host_phys, g_iso_size);
     hype_ahci_reset(&g_iso_2_ahci);
 
     cmd_list_phys = (uint64_t)(uintptr_t)g_iso_2_cmd_list;
